@@ -1,11 +1,8 @@
-import Vue from 'vue'
-import './plugins/vuetify'
+import './index.css'
+import { createApp } from 'vue'
+import { createPinia } from 'pinia'
 import App from './App.vue'
-import Clipboard from 'v-clipboard'
 
-Vue.config.productionTip = false
-Vue.use(Clipboard)
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+const app = createApp(App)
+app.use(createPinia())
+app.mount('#app')
