@@ -1,63 +1,94 @@
 <template>
-	<v-app dark>
-		<v-toolbar app flat>
-			<v-toolbar-title class="headline text-uppercase">
-				<span class="font-weight-light">Cugetarile </span>
-				<span class="font-weight-black">Tribunului</span>
-			</v-toolbar-title>
-			<v-spacer></v-spacer>
-			<v-btn
-					depressed
-					flat
-					href="https://github.com/abaicus/cugetarile-tribunului"
-					target="_blank"
-			>
-				<v-icon>star</v-icon>
-			</v-btn>
-			<v-btn color="info" @click="pause()">
-				<v-icon>pause</v-icon>
-			</v-btn>
-		</v-toolbar>
-		<v-content>
-			<Main/>
-		</v-content>
-		<v-footer dark>
-			<v-card width="100%">
-				<v-card-actions>
-					<v-spacer></v-spacer>
-					<div>Powered by <a href="https://vuetifyjs.com" target="_blank">Vuetify</a> & <a
-							href="https://vuejs.org/" target="_blank">VueJS</a></div>
-					<div>&nbsp;&copy; {{ new Date().getFullYear() }}</div>
-				</v-card-actions>
-			</v-card>
-		</v-footer>
-	</v-app>
+  <div class="flex flex-col min-h-screen bg-black/95 text-white antialiased">
+    <!-- Header -->
+    <header class="bg-zinc-950 sticky top-0 z-50 w-full">
+      <nav class="container mx-auto w-full px-4 py-3 flex justify-between items-center">
+        <h1 class="text-xl flex items-center gap-2 uppercase">
+          <span class="font-light">Cugetarile</span>
+          <span class="font-bold">Tribunului</span>
+        </h1>
+        <div class="ml-auto flex items-center gap-3 space-x-2">
+          <a href="https://buymeacoffee.com/abaicus" target="_blank"
+            class="p-2 rounded-sm hover:bg-neutral-800 border-neutral-900 border transition-colors">
+            <BMCLogo />
+          </a>
+          <a href="https://github.com/abaicus/cugetarile-tribunului" target="_blank"
+            class="p-2 rounded-sm hover:bg-neutral-800 border-neutral-900 border transition-colors hidden md:block">
+            <svg class="w-6 h-6" fill="currentColor" viewBox="0 0 24 24" aria-hidden="true">
+              <path fill-rule="evenodd"
+                d="M12 2C6.477 2 2 6.484 2 12.017c0 4.425 2.865 8.18 6.839 9.504.5.092.682-.217.682-.483 0-.237-.008-.868-.013-1.703-2.782.605-3.369-1.343-3.369-1.343-.454-1.158-1.11-1.466-1.11-1.466-.908-.62.069-.608.069-.608 1.003.07 1.531 1.032 1.531 1.032.892 1.53 2.341 1.088 2.91.832.092-.647.35-1.088.636-1.338-2.22-.253-4.555-1.113-4.555-4.951 0-1.093.39-1.988 1.029-2.688-.103-.253-.446-1.272.098-2.65 0 0 .84-.27 2.75 1.026A9.564 9.564 0 0112 6.844c.85.004 1.705.115 2.504.337 1.909-1.296 2.747-1.027 2.747-1.027.546 1.379.202 2.398.1 2.651.64.7 1.028 1.595 1.028 2.688 0 3.848-2.339 4.695-4.566 4.943.359.309.678.92.678 1.855 0 1.338-.012 2.419-.012 2.747 0 .268.18.58.688.482A10.019 10.019 0 0022 12.017C22 6.484 17.522 2 12 2z"
+                clip-rule="evenodd" />
+            </svg>
+          </a>
+          <button @click="togglePlayPause" :disabled="!isAudioAvailable || !isPlaying" :class="[
+            'p-2 rounded-sm transition-colors focus:outline-none focus:border-indigo-300 focus:ring-indigo-300 focus:ring-1 border hidden md:block',
+            isAudioAvailable && isPlaying
+              ? 'bg-indigo-600 hover:bg-indigo-700'
+              : 'bg-indigo-600 opacity-50 cursor-not-allowed'
+          ]">
+            <PauseIcon v-if="isPlaying" class="w-6 h-6" />
+            <PlayIcon v-else class="w-6 h-6" />
+          </button>
+        </div>
+      </nav>
+    </header>
+
+
+    <!-- Main Content -->
+    <main class="grow place-items-center grid w-full px-4 py-8">
+      <Main />
+    </main>
+
+    <!-- Footer -->
+    <footer class="bg-zinc-950 mt-auto font-mono">
+      <div class="container mx-auto px-4 py-3 text-center flex justify-between font-light text-xs">
+
+        <div class="flex items-center gap-1">
+          <span>Powered by</span>
+          <a href="https://vuejs.org/" target="_blank" class="hover:opacity-50">
+            <span class="sr-only">
+              Vue.js
+            </span>
+            <svg xmlns="http://www.w3.org/2000/svg" class="h-4 fill-current" viewBox="0 0 24 24">
+              <path d="m12 12.765 5.592-9.437h-3.276L12 7.33v.002L9.688 3.328h-3.28z" />
+              <path d="M18.461 3.332 12 14.235 5.539 3.332H1.992L12 20.672l10.008-17.34z" />
+            </svg>
+          </a>
+          <span>|</span>
+          <a href="https://tailwindcss.com" target="_blank" class="hover:opacity-50">
+            <span class="sr-only">Tailwind CSS</span>
+            <svg viewBox="0 0 248 31" class="h-3 fill-current">
+              <path
+                d="M25.517 0C18.712 0 14.46 3.382 12.758 10.146c2.552-3.382 5.529-4.65 8.931-3.805 1.941.482 3.329 1.882 4.864 3.432 2.502 2.524 5.398 5.445 11.722 5.445 6.804 0 11.057-3.382 12.758-10.145-2.551 3.382-5.528 4.65-8.93 3.804-1.942-.482-3.33-1.882-4.865-3.431C34.736 2.92 31.841 0 25.517 0zM12.758 15.218C5.954 15.218 1.701 18.6 0 25.364c2.552-3.382 5.529-4.65 8.93-3.805 1.942.482 3.33 1.882 4.865 3.432 2.502 2.524 5.397 5.445 11.722 5.445 6.804 0 11.057-3.381 12.758-10.145-2.552 3.382-5.529 4.65-8.931 3.805-1.941-.483-3.329-1.883-4.864-3.432-2.502-2.524-5.398-5.446-11.722-5.446z">
+              </path>
+              <path
+                d="M76.546 12.825h-4.453v8.567c0 2.285 1.508 2.249 4.453 2.106v3.463c-5.962.714-8.332-.928-8.332-5.569v-8.567H64.91V9.112h3.304V4.318l3.879-1.143v5.937h4.453v3.713zM93.52 9.112h3.878v17.849h-3.878v-2.57c-1.365 1.891-3.484 3.034-6.285 3.034-4.884 0-8.942-4.105-8.942-9.389 0-5.318 4.058-9.388 8.942-9.388 2.801 0 4.92 1.142 6.285 2.999V9.112zm-5.674 14.636c3.232 0 5.674-2.392 5.674-5.712s-2.442-5.711-5.674-5.711-5.674 2.392-5.674 5.711c0 3.32 2.442 5.712 5.674 5.712zm16.016-17.313c-1.364 0-2.477-1.142-2.477-2.463a2.475 2.475 0 012.477-2.463 2.475 2.475 0 012.478 2.463c0 1.32-1.113 2.463-2.478 2.463zm-1.939 20.526V9.112h3.879v17.849h-3.879zm8.368 0V.9h3.878v26.06h-3.878zm29.053-17.849h4.094l-5.638 17.849h-3.807l-3.735-12.03-3.771 12.03h-3.806l-5.639-17.849h4.094l3.484 12.315 3.771-12.315h3.699l3.734 12.315 3.52-12.315zm8.906-2.677c-1.365 0-2.478-1.142-2.478-2.463a2.475 2.475 0 012.478-2.463 2.475 2.475 0 012.478 2.463c0 1.32-1.113 2.463-2.478 2.463zm-1.939 20.526V9.112h3.878v17.849h-3.878zm17.812-18.313c4.022 0 6.895 2.713 6.895 7.354V26.96h-3.878V16.394c0-2.713-1.58-4.14-4.022-4.14-2.55 0-4.561 1.499-4.561 5.14v9.567h-3.879V9.112h3.879v2.285c1.185-1.856 3.124-2.749 5.566-2.749zm25.282-6.675h3.879V26.96h-3.879v-2.57c-1.364 1.892-3.483 3.034-6.284 3.034-4.884 0-8.942-4.105-8.942-9.389 0-5.318 4.058-9.388 8.942-9.388 2.801 0 4.92 1.142 6.284 2.999V1.973zm-5.674 21.775c3.232 0 5.674-2.392 5.674-5.712s-2.442-5.711-5.674-5.711-5.674 2.392-5.674 5.711c0 3.32 2.442 5.712 5.674 5.712zm22.553 3.677c-5.423 0-9.481-4.105-9.481-9.389 0-5.318 4.058-9.388 9.481-9.388 3.519 0 6.572 1.82 8.008 4.605l-3.34 1.928c-.79-1.678-2.549-2.749-4.704-2.749-3.16 0-5.566 2.392-5.566 5.604 0 3.213 2.406 5.605 5.566 5.605 2.155 0 3.914-1.107 4.776-2.749l3.34 1.892c-1.508 2.82-4.561 4.64-8.08 4.64zm14.472-13.387c0 3.249 9.661 1.285 9.661 7.89 0 3.57-3.125 5.497-7.003 5.497-3.591 0-6.177-1.607-7.326-4.177l3.34-1.927c.574 1.606 2.011 2.57 3.986 2.57 1.724 0 3.052-.571 3.052-2 0-3.176-9.66-1.391-9.66-7.781 0-3.356 2.909-5.462 6.572-5.462 2.945 0 5.387 1.357 6.644 3.713l-3.268 1.82c-.647-1.392-1.904-2.035-3.376-2.035-1.401 0-2.622.607-2.622 1.892zm16.556 0c0 3.249 9.66 1.285 9.66 7.89 0 3.57-3.124 5.497-7.003 5.497-3.591 0-6.176-1.607-7.326-4.177l3.34-1.927c.575 1.606 2.011 2.57 3.986 2.57 1.724 0 3.053-.571 3.053-2 0-3.176-9.66-1.391-9.66-7.781 0-3.356 2.908-5.462 6.572-5.462 2.944 0 5.386 1.357 6.643 3.713l-3.268 1.82c-.646-1.392-1.903-2.035-3.375-2.035-1.401 0-2.622.607-2.622 1.892z">
+              </path>
+            </svg>
+          </a>
+        </div>
+
+        <p>&copy; {{ new Date().getFullYear() }}</p>
+      </div>
+    </footer>
+  </div>
 </template>
 
-<script>
-	import Main from '@/components/Main';
+<script setup>
+import { computed } from 'vue';
+import { PauseIcon, PlayIcon } from '@heroicons/vue/20/solid';
+import { useAudioPlayerStore } from './stores/audioPlayerStore';
+import Main from './components/Main.vue';
+import BMCLogo from './components/BMCLogo.vue'
 
-	export default {
-		name: 'App',
-		components: {
-			Main
-		},
-		methods: {
-			pause: function() {
-				if ( window.currentlyPlaying.audio === null ) {
-					return false;
-				}
+const store = useAudioPlayerStore();
 
-				window.currentlyPlaying.audio.pause();
-				window.currentlyPlaying.component.playing = false;
-			}
-		}
-	};
+const isPlaying = computed(() => store.isPlaying);
+const isAudioAvailable = computed(() => store.currentAudio !== null);
+
+const togglePlayPause = () => {
+  if (isAudioAvailable.value) {
+    store.togglePlayPause();
+  }
+};
 </script>
-<style lang="scss">
-	footer a {
-		color: #fff;
-		text-decoration: none;
-		font-weight: 800;
-	}
-</style>
