@@ -1,28 +1,20 @@
 <template>
-  <div>
-    <v-container>
-      <v-layout text-xs-center wrap>
-      <v-flex xs12>
-          <v-card light>
-            <v-avatar size="100" class="mt-3">
-              <v-img :src="require('../assets/avatar.jpg')"></v-img>
-            </v-avatar>
-          <sound-board></sound-board>
-          </v-card>
-        </v-flex>
-      </v-layout>
-    </v-container>
+  <div class="px-6 md:p-6 w-full text-center">
+    <div class="h-28 w-28 mx-auto mb-5 md:mb-10">
+      <img src="../assets/avatar.jpg" alt="Avatar"
+        class="rounded-full object-cover w-full h-full filter hue-rotate-30" />
+    </div>
+    <SoundBoard />
   </div>
 </template>
 
 <script>
-import SoundBoard from "@/components/SoundBoard";
+import SoundBoard from "./SoundBoard.vue";
+
 export default {
+  name: 'MainComponent',
   components: {
     SoundBoard
-  },
+  }
 };
 </script>
-
-<style>
-</style>
